Renaming: from SovlerConfiguration to SolverAdapterInfo.
