+++ src/checkers/inference/solver/backend/logiql/LogiqlSolverInfo.java
@@ -2,9 +2,9 @@ package checkers.inference.solver.backend.logiql;
 
 import checkers.inference.solver.backend.FormatTranslator;
 import checkers.inference.solver.backend.SolverAdapter;
-import checkers.inference.solver.backend.SolverConfiguration;
+import checkers.inference.solver.backend.SolverAdapterInfo;
 
-public class LogiqlSolverConfiguration implements SolverConfiguration {
+public class LogiqlSolverInfo implements SolverAdapterInfo {
 
     @Override
     public Class<? extends SolverAdapter<?>> getSolverAdapterClass() {
