+++ src/checkers/inference/solver/backend/z3/Z3SolverInfo.java
@@ -2,9 +2,9 @@ package checkers.inference.solver.backend.z3;
 
 import checkers.inference.solver.backend.FormatTranslator;
 import checkers.inference.solver.backend.SolverAdapter;
-import checkers.inference.solver.backend.SolverConfiguration;
+import checkers.inference.solver.backend.SolverAdapterInfo;
 
-public class Z3SolverConfiguration implements SolverConfiguration {
+public class Z3SolverInfo implements SolverAdapterInfo {
     @Override
     public Class<? extends SolverAdapter<?>> getSolverAdapterClass() {
         return Z3Solver.class;
